+++ NEWS.md
@@ -1,10 +1,11 @@
 # odbc (development version)
 
+* `databricks()` now works with manually supplied `pwd` and `uid` (#690).
+
 * Added a function `odbcListConfig()` to help locate configuration files on
   macOS and Linux (@simonpcouch, #565).
 
 * Use correct parent class for Oracle (#685).
-
 
 # odbc 1.4.0